requirements.txt
legacy-cgi
Django==3.1.1
virtualenv==20.0.31
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
Pillow
requests
google-generativeai
python-dotenv
